Reallocate unit test
